backend/requirements copy.txt
flask
flask_cors
influxdb-client
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
